refactor: 베타 기간 뱃지 획득 책임 분리
