+++ src/main/java/develop/grassserver/member/badge/BadgeService.java
@@ -1,5 +1,6 @@
 package develop.grassserver.member.badge;
 
+import develop.grassserver.member.Member;
 import develop.grassserver.member.badge.dto.FindAllMemberBadgesResponse;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -16,5 +17,18 @@ public class BadgeService {
     public FindAllMemberBadgesResponse findAllMemberBadges(Long memberId) {
         List<MemberBadge> badges = memberBadgeRepository.findAllByMemberId(memberId);
         return FindAllMemberBadgesResponse.from(badges);
+    }
+
+    @Transactional
+    public void saveBetaMemberBadge(Member member) {
+        MemberBadge betaJoinBadge = createBetaJoinBadge(member);
+        memberBadgeRepository.save(betaJoinBadge);
+    }
+
+    private MemberBadge createBetaJoinBadge(Member member) {
+        return MemberBadge.builder()
+                .badge(Badge.EARLY_ADOPTER)
+                .member(member)
+                .build();
     }
 }
